requirements.txt
Django==4.2.15
gunicorn==23.0.0
django-cors-headers==4.3.1
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
psycopg2-binary==2.9.5
djangorestframework==3.16.0
celery===5.4.0
django-celery-beat==2.6.0
django-celery-results==2.4.0
flower===2.0.1
redis==6.2.0
django-environ==0.10.0
requests==2.32.4
